+++ R/classes.R
@@ -4,27 +4,23 @@
 #' @title NULL or char class union
 #' @description class to allow either NULL or character
 #' @keywords internal
+#' @noRd
 setClassUnion('nullOrChar', c('NULL', 'character'))
 
 #' @title NULL or list class union
 #' @description class to allow either NULL or list
 #' @keywords internal
+#' @noRd
 setClassUnion('nullOrList', c('NULL', 'list'))
 
 #' @title NULL or data.table class union
 #' @description class to allow either NULL or data.table
 #' @keywords internal
+#' @noRd
+setOldClass('data.table')
 setClassUnion('nullOrDatatable', c('NULL', 'data.table'))
 
-## * Define external classes ####
 
-#' data.table S4 class for method dispatch
-#' @name data.table-class
-#' @aliases data.table
-#' @family data.table
-#'
-#' @exportClass data.table
-setOldClass('data.table')
 
 
 
@@ -34,7 +30,8 @@ setOldClass('data.table')
 # VIRTUAL CLASSES ####
 
 # ** nameData Class ####
-#'
+#' @keywords internal
+#' @noRd
 setClass('nameData',
          contains = 'VIRTUAL',
          slots = list(name = 'character'),
@@ -42,7 +39,8 @@ setClass('nameData',
 
 # ** exprData Class ####
 #' Basic class for classes with expression information
-#'
+#' @keywords internal
+#' @noRd
 setClass('exprData',
          contains = 'VIRTUAL',
          slots = list(exprMat = 'ANY'),
@@ -57,6 +55,8 @@ setClass('exprData',
 #' information is stored within data.table objects and should work similarly to
 #' data.table when interacting with some basic generic operators for data
 #' retreival and setting.
+#' @keywords internal
+#' @noRd
 setClass('coordDataDT',
          contains = 'VIRTUAL',
          slots = list(coordinates = 'data.table'),
@@ -86,6 +86,8 @@ setMethod('initialize', 'coordDataDT',
 #' Classes that inherit from this class will contain a metadata slot that stores
 #' information in a data.table and should work similarly to data.table when interacting
 #' with some basic generic operators for data retrieval and setting
+#' @keywords internal
+#' @noRd
 setClass('metaData',
          contains = 'VIRTUAL',
          slots = list(metaDT = 'data.table',
@@ -107,6 +109,8 @@ setMethod('initialize', 'metaData',
 
 # ** enrData ####
 #' enrData
+#' @keywords internal
+#' @noRd
 setClass('enrData',
          contains = 'VIRTUAL',
          slots = list(method = 'character',
@@ -127,6 +131,8 @@ setMethod('initialize', 'enrData',
 
 
 # ** nnData ####
+#' @keywords internal
+#' @noRd
 setClass('nnData',
          contains = 'VIRTUAL',
          slots = list(nn_type = 'character',
@@ -136,6 +142,8 @@ setClass('nnData',
 
 
 # ** spatNetData ####
+#' @keywords internal
+#' @noRd
 setClass('spatNetData',
          contains = 'VIRTUAL',
          slots = list(method = 'character',
@@ -166,6 +174,8 @@ setMethod('initialize', 'spatNetData',
 
 
 # ** spatGridData ####
+#' @keywords internal
+#' @noRd
 setClass('spatGridData',
          contains = 'VIRTUAL',
          slots = list(method = 'character',
@@ -196,6 +206,8 @@ setMethod('initialize', 'spatGridData',
 #' information and polygons that are provided as multiple z layers. Provenance
 #' is Giotto's method of mapping this aggregated information back to the original
 #' z layers that were used in its generation.
+#' @keywords internal
+#' @noRd
 setClass('provData',
          contains = 'VIRTUAL',
          slots = list(provenance = 'ANY'),
@@ -211,7 +223,8 @@ setClass('provData',
 #' Subcellular information such as poly data in \code{spatial_info} slot essentially define their
 #' own spatial units. Within slots that deal with classes that contain spatData,
 #' there is a nesting structure that first nests by spatial unit.
-#'
+#' @keywords internal
+#' @noRd
 setClass('spatData',
          contains = c('provData', 'VIRTUAL'),
          slots = list(spat_unit = 'character'), # not allowed to be NULL
@@ -229,7 +242,8 @@ setClass('spatData',
 #' which feature type the data is. Within slots that deal with classes that contain
 #' featData, there is a nesting structure that usually first nests by spatial unit
 #' and then by feature type
-#'
+#' @keywords internal
+#' @noRd
 setClass('featData',
          contains = 'VIRTUAL',
          slots = list(feat_type = 'character'), # not allowed to be NULL
@@ -260,6 +274,8 @@ setClass('miscData',
 
 # ** spatFeatData ####
 #' Superclass for classes that contain both spatial and feature data
+#' @keywords internal
+#' @noRd
 setClass('spatFeatData',
          contains = c('spatData', 'featData', 'VIRTUAL'))
 
@@ -362,6 +378,7 @@ setClass('spatFeatData',
 #' Supported updates:
 #' \itemize{
 #'   \item{3.2.0 update adding multiomics slot}
+#'   \item{3.4.0 update adding dbInfo slot}
 #'   \item{master branch to suite - TODO}
 #' }
 #' @examples
@@ -379,6 +396,12 @@ updateGiottoObject = function(gobject) {
   if(is.null(attr(gobject, 'multiomics'))) {
     attr(gobject, 'multiomics') = NA
     gobject@multiomics = NULL
+  }
+
+  # 3.4.0 release adds dbInfo slot
+  if(is.null(attr(gobject, 'dbInfo'))) {
+    attr(gobject, 'dbInfo') = NA
+    gobject@dbInfo = NULL
   }
 
   return(gobject)
@@ -411,6 +434,7 @@ updateGiottoObject = function(gobject) {
 #' @slot largeImages slot to store giottoLargeImage objects
 #' @slot parameters slot to save parameters that have been used
 #' @slot instructions slot for global function instructions
+#' @slot dbInfo slot for database backend
 #' @slot offset_file offset file used to stitch together image fields
 #' @slot OS_platform Operating System to run Giotto analysis on
 #' @slot join_info information about joined Giotto objects
@@ -444,6 +468,7 @@ giotto <- setClass(
     largeImages = "ANY",
     parameters = "ANY",
     instructions = "ANY",
+    dbInfo = 'ANY',
     offset_file = "ANY",
     OS_platform = "ANY",
     join_info = "ANY",
@@ -470,6 +495,7 @@ giotto <- setClass(
     largeImages = NULL,
     parameters = list(),
     instructions = NULL,
+    dbInfo = NULL,
     offset_file = NULL,
     OS_platform = NULL,
     join_info = NULL,
@@ -678,7 +704,7 @@ setMethod('initialize', signature('giotto'), function(.Object, ...) {
       list_match = sapply(exp_list_names, setequal, exp_list_names[[1L]])
       if(!all(list_match)) {
         print(list_match)
-        warning(wrap_text(
+        warning(wrap_txt(
           'spat_unit:', unique_expr_sets$spat_unit[[set_i]], '/',
           'feat_type:', unique_expr_sets$feat_type[[set_i]],
           '\nNot all expression matrices share the same cell_IDs'
@@ -2467,22 +2493,54 @@ setMethod(
 #' @param spat_unit spatial unit of expression (e.g. 'cell')
 #' @param feat_type feature type of expression (e.g. 'rna', 'protein')
 #' @param provenance origin data of expression information (if applicable)
+#' @param matrix_type type of matrix to use
+#' @param cores number of cores to use
+#' @param remote_dir (optional) directory of remote backend. Default is tempdir()
+#' @param db_extension (optional) database file extension
+#' @param ... additional params to pass to matrix reading functions
 #' @param misc misc
 createExprObj = function(expression_data,
                          name = 'test',
                          spat_unit = 'cell',
                          feat_type = 'rna',
                          provenance = NULL,
-                         misc = NULL) {
+                         expression_matrix_class = c('Matrix', 'HDF5Matrix', 'dbMatrix'),
+                         sparse = TRUE,
+                         cores = determine_cores(),
+                         remote_dir = ':temp:',
+                         db_extension = '.duckdb',
+                         misc = NULL,
+                         ...) {
+  expression_matrix_class = match.arg(
+    expression_matrix_class, choices = c('Matrix', 'HDF5Matrix', 'dbMatrix')
+  )
+  # Backends: check package and establish remote_dir
+  switch(expression_matrix_class,
+         'dbMatrix' = {
+           package_check('GiottoDB', repository = 'github', github_repo = 'drieslab/GiottoDB')
+           remote_dir = GiottoDB::getDBPath(path = remote_dir, extension = db_extension)
+         },
+         'DelayedArray' = {
+           package_check('HDF5Array', repository = 'Bioc')
+           remote_dir
+           # TODO figure out remote directory/h5 folder
+         })
 
-  exprMat = evaluate_expr_matrix(expression_data)
+
+  exprMat = evaluate_expr_matrix(inputmatrix = expression_data,
+                                 sparse = sparse,
+                                 expression_matrix_class = expression_matrix_class,
+                                 remote_dir = remote_dir,
+                                 remote_name = paste(spat_unit, feat_type, name, sep = '_'),
+                                 cores = cores,
+                                 ...)
 
   create_expr_obj(name = name,
                   exprMat = exprMat,
                   spat_unit = spat_unit,
                   feat_type = feat_type,
                   provenance = provenance,
-                  misc = misc)
+                  misc = list('matrix_type' = expression_matrix_class))
 }